// src/types/customDataTypes.ts
interface contentObjectTypes {
  id: number;
  imgSrc?: string;
  title?: string;
  btnText?: string;
  path?: string;
  boldItem?: boldItemTypes;
  paraContent?: paraContentTypes[];
  listItem?: listItemTypes[];
  listItemWithTitle?: listItemWithTitleTypes[];
  subListItem?: subListItemTypes[];
  tableContent?: customLibraryDataType[];
  cardData?: cardTypesData[];
  accordianShow?: boolean;
  tableShow?: boolean;
  content?: string;
  accordianData?: accordianDataType[];
  cardItem?: cardItemType[];
  galleryData?: galleryDataTypes[];
  vissionData?: vissionDataTypes[];
  awardsData?: awardsDataTypes[];
  contentData?: contentDataTypes[];

}

export interface cardContentChildTypes {
  id: number;
  imgSrc: string;
  title: string;
  date: string;
  btnTitle: string;
  path?: string;
}
export interface paraContentChildTypes {
  id: number;
  value: string;
  bold?: string;
}
export interface contentTypes {
  id: number;
  title: string;
  paraContentChild: paraContentChildTypes[];
  cardContentChild?: cardContentChildTypes[];
}

interface LaboratoriesContentTypes {
  id?: number;
  para1?: string;
  subtitle?: string;
  paraContentChild?: paraContentChildTypes[];
}

interface contentCardTypes {
  id: number;
  title?: string;
  imgSrc?: string;
  paraContent?: paraContentTypes[];
  LaboratoriesContent?: LaboratoriesContentTypes[];
  laboratory_desc?: string;
  subtitle?: string;
  paraContentChild?: paraContentChildTypes[];
}

interface contetWithBoldTypes {
  id: 0;
  boldValue: string;
  value: string;
}
interface contentPdfDataTypes {
  id: number;
  value?: string;
  path?: string;
  boldValue?: string;
  imgsrc1?: string;
  imgsrc2?: string;
  contetWithBold?: contetWithBoldTypes[];
}
interface contentPdfObjectTypes {
  id: number;
  subtitle?: string;
  contentPdfData: contentPdfDataTypes[];
}
interface contentWithPdfTypes {
  id: number;
  title?: string;
  contentPdfObject: contentPdfObjectTypes;
}
interface tabDataTypes {
  id: number;
  tab: string;
  content?: contentTypes[];
  isMode?: boolean;
  contentObject?: contentObjectTypes;
  contentCard?: contentCardTypes[];
  contentWithPdf?: contentWithPdfTypes[];
  isLesscontent?: boolean;
}
export interface customTabDataTypes {
  id: number;
  isAnualFast?: boolean;
  tabData: tabDataTypes[];
  isDepartment?: boolean;
}